refactoring scSheet types
